--- a/fe/fe-core/src/main/java/com/starrocks/sql/optimizer/operator/logical/LogicalAggregationOperator.java
+++ b/fe/fe-core/src/main/java/com/starrocks/sql/optimizer/operator/logical/LogicalAggregationOperator.java
@@ -319,6 +319,11 @@ public class LogicalAggregationOperator extends LogicalOperator {
             return this;
         }
 
+        public Builder setSplit(boolean split) {
+            builder.isSplit = split;
+            return this;
+        }
+
         public Builder setPartitionByColumns(
                 List<ColumnRefOperator> partitionByColumns) {
             builder.partitionByColumns = partitionByColumns;