--- a/fe/fe-core/src/main/java/com/starrocks/sql/optimizer/operator/logical/LogicalRepeatOperator.java
+++ b/fe/fe-core/src/main/java/com/starrocks/sql/optimizer/operator/logical/LogicalRepeatOperator.java
@@ -33,6 +33,7 @@ public class LogicalRepeatOperator extends LogicalOperator {
     private List<ColumnRefOperator> outputGrouping;
     private List<List<ColumnRefOperator>> repeatColumnRefList;
     private List<List<Long>> groupingIds;
+    private boolean hasPushDown;
 
     public LogicalRepeatOperator(List<ColumnRefOperator> outputGrouping,
                                  List<List<ColumnRefOperator>> repeatColumnRefList,
@@ -41,6 +42,7 @@ public class LogicalRepeatOperator extends LogicalOperator {
         this.outputGrouping = outputGrouping;
         this.repeatColumnRefList = repeatColumnRefList;
         this.groupingIds = groupingIds;
+        this.hasPushDown = false;
     }
 
     private LogicalRepeatOperator() {
@@ -59,6 +61,10 @@ public class LogicalRepeatOperator extends LogicalOperator {
         return groupingIds;
     }
 
+    public boolean hasPushDown() {
+        return hasPushDown;
+    }
+
     @Override
     public ColumnRefSet getOutputColumns(ExpressionContext expressionContext) {
         ColumnRefSet outputColumns = new ColumnRefSet(outputGrouping);
@@ -72,7 +78,7 @@ public class LogicalRepeatOperator extends LogicalOperator {
     @Override
     public RowOutputInfo deriveRowOutputInfo(List<OptExpression> inputs) {
         List<ColumnOutputInfo> columnOutputInfoList = Lists.newArrayList();
-        outputGrouping.stream().forEach(e -> columnOutputInfoList.add(new ColumnOutputInfo(e, e)));
+        outputGrouping.forEach(e -> columnOutputInfoList.add(new ColumnOutputInfo(e, e)));
         for (ColumnOutputInfo columnOutputInfo : inputs.get(0).getRowOutputInfo().getColumnOutputInfo()) {
             columnOutputInfoList.add(new ColumnOutputInfo(columnOutputInfo.getColumnRef(), columnOutputInfo.getColumnRef()));
         }
@@ -102,14 +108,18 @@ public class LogicalRepeatOperator extends LogicalOperator {
         LogicalRepeatOperator that = (LogicalRepeatOperator) o;
         return Objects.equals(outputGrouping, that.outputGrouping) &&
                 Objects.equals(repeatColumnRefList, that.repeatColumnRefList) &&
-                Objects.equals(groupingIds, that.groupingIds);
+                Objects.equals(groupingIds, that.groupingIds) &&
+                Objects.equals(hasPushDown, that.hasPushDown);
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(super.hashCode(), outputGrouping, repeatColumnRefList);
+        return Objects.hash(super.hashCode(), outputGrouping, repeatColumnRefList, hasPushDown);
     }
 
+    public static Builder builder() {
+        return new Builder();
+    }
     public static class Builder
             extends LogicalOperator.Builder<LogicalRepeatOperator, LogicalRepeatOperator.Builder> {
 
@@ -118,8 +128,18 @@ public class LogicalRepeatOperator extends LogicalOperator {
             return new LogicalRepeatOperator();
         }
 
+        public LogicalRepeatOperator.Builder setHasPushDown(boolean hasPushDown) {
+            builder.hasPushDown = hasPushDown;
+            return this;
+        }
+
         public LogicalRepeatOperator.Builder setOutputGrouping(List<ColumnRefOperator> outputGrouping) {
             builder.outputGrouping = outputGrouping;
+            return this;
+        }
+
+        public LogicalRepeatOperator.Builder setGroupingIds(List<List<Long>> groupingIds) {
+            builder.groupingIds = groupingIds;
             return this;
         }