--- a/fe/fe-core/src/test/java/com/starrocks/connector/parser/trino/TrinoTestBase.java
+++ b/fe/fe-core/src/test/java/com/starrocks/connector/parser/trino/TrinoTestBase.java
@@ -147,6 +147,7 @@ public class TrinoTestBase {
         FeConstants.runningUnitTest = false;
 
         connectContext.getSessionVariable().setSqlDialect("trino");
+        connectContext.getSessionVariable().setCboPushDownGroupingSet(false);
     }
 
     public static StatementBase analyzeSuccess(String originStmt) {