… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rule/RuleType.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/transformer/QueryTransformer.java
… fe/fe-core/src/test/java/com/starrocks/sql/plan/GroupingSetsTest.java
… fe/fe-core/src/test/java/com/starrocks/sql/plan/PlanTestBase.java
… fe/fe-core/src/test/java/com/starrocks/sql/plan/TPCDS1TTest.java