## Why I'm doing:

## What I'm doing:
grouping set sql
```
     select a, b, c, d, e sum(f) from t group by rollup(a, b, c, d, e);
```

rewrite to:
```
   with cte1 as (select a, b, c, d, e, sum(f) x from t group by a, b, c, d, e -- shuffle by a high cardinality column, for better effect on local aggregate
     selec * fom cte1
     union all
     select a, b, c, d, null, sum(x) x from cte1 group by rollup(a, b, c, d)
```


if the last column in rollup/cube/groupingset：
* is a high cardinality column,  it's has a significant performance improvement. (like tpcds q67)
* is a low cardinality column,  it's has no effect.

vs `set enable_rewrite_groupingsets_to_union_all = true` (will rewite whole grouping set to CTE)
* save CTE copy cost
* I think has no any particular disadvantages, maybe repeat node will use more memory?

tpcds Q67 100g: 13.492s -> 10.253s
tpcds Q67 1t: 29.06s -> 10s

Fixes #issue

## What type of PR is this:

- [ ] BugFix
- [ ] Feature
- [x] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

Does this PR entail a change in behavior?

- [ ] Yes, this PR will result in a change in behavior.
- [x] No, this PR will not result in a change in behavior.

If yes, please specify the type of change:

- [ ] Interface/UI changes: syntax, type conversion, expression evaluation, display information
- [ ] Parameter changes: default values, similar parameters but with different default values
- [ ] Policy changes: use new policy to replace old one, functionality automatically enabled
- [ ] Feature removed
- [ ] Miscellaneous: upgrade & downgrade compatibility, etc.

## Checklist:

- [x] I have added test cases for my bug fix or my new feature
- [ ] This pr needs user documentation (for new or modified features or behaviors)
  - [ ] I have added documentation for my new feature or new function
- [ ] This is a backport pr

## Bugfix cherry-pick branch check:
- [x] I have checked the version labels which the pr will be auto-backported to the target branch
  - [ ] 3.3
  - [ ] 3.2
  - [ ] 3.1
  - [ ] 3.0
  - [ ] 2.5